refactor: VisualTransformation OffsetMapping 분리
