+++ app/src/main/java/woowacourse/payments/ui/newcard/util/transformation/ExpirationDateVisualTransformation.kt
@@ -7,32 +7,18 @@ import androidx.compose.ui.text.input.VisualTransformation
 
 class ExpirationDateVisualTransformation : VisualTransformation {
     override fun filter(text: AnnotatedString): TransformedText {
-        val trimmed = text.text.filter { it.isDigit() }.take(4)
-
-        val formattedText = trimmed.chunked(2).joinToString(" / ")
-
-        val offsetTranslator =
-            object : OffsetMapping {
-                override fun originalToTransformed(offset: Int): Int {
-                    if (offset <= 0) return 0
-                    if (offset >= trimmed.length) return formattedText.length
-
-                    val sepCount = if (offset > 2) 1 else 0
-                    return offset + sepCount * 3
-                }
-
-                override fun transformedToOriginal(offset: Int): Int {
-                    if (offset <= 0) return 0
-                    if (offset >= formattedText.length) return trimmed.length
-
-                    val sepCount = if (offset > 2 + 3) 1 else 0
-                    return (offset - sepCount * 3).coerceIn(0, trimmed.length)
-                }
-            }
+        val trimmed = text.text.filter { it.isDigit() }.take(MAX_LENGTH)
+        val formattedText = trimmed.chunked(CHUNK_SIZE).joinToString(SEPARATOR)
 
         return TransformedText(
             AnnotatedString(formattedText),
-            offsetTranslator,
+            ExpirationDateOffsetMapping(trimmed, formattedText),
         )
+    }
+
+    companion object {
+        private const val MAX_LENGTH = 4
+        private const val CHUNK_SIZE = 2
+        private const val SEPARATOR = " / "
     }
 }
